// src/api/generated/schemas/validateProviderResponse.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */
import type { ValidationCheckDto } from './validationCheckDto';

export interface ValidateProviderResponse {
  isValid?: boolean;
  /** @nullable */
  message?: string | null;
  checks?: ValidationCheckDto[];
}